feat: Reuse the salesforce hook for sync
