+++ src/elements/fields/DropdownField.tsx
@@ -8,7 +8,7 @@ import { getStateOptions, hasState } from '../components/data/states';
 import { css, Global } from '@emotion/react';
 import { hoverStylesGuard, iosScrollOnFocus } from '../../utils/browser';
 import { fieldValues } from '../../utils/init';
-import FeatheryClient from '../../utils/featheryClient';
+import useSalesforceSync from '../../hooks/useSalesforceSync';
 
 export default function DropdownField({
   element,
@@ -36,8 +36,9 @@ export default function DropdownField({
   const containerRef = useRef(null);
   const servar = element.servar;
   const short = servar.metadata.store_abbreviation;
-  const [dynamicOptions, setDynamicOptions] = useState<string[]>([]);
-  const [loadingDynamicOptions, setLoadingDynamicOptions] = useState(false);
+  const { dynamicOptions, loadingDynamicOptions } = useSalesforceSync(
+    servar.metadata.salesforce_sync
+  );
 
   useEffect(() => {
     if (servar.type === 'gmap_state') {
@@ -49,27 +50,6 @@ export default function DropdownField({
       setCurCountry(code || 'us');
     }
   }, [countryCode, setCurCountry]);
-
-  useEffect(() => {
-    const salesforceSync = servar.metadata.salesforce_sync;
-    if (!salesforceSync) return;
-
-    const fetchSalesforceOptions = async () => {
-      setLoadingDynamicOptions(true);
-      try {
-        const client = new FeatheryClient();
-        const data = await client.fetchSalesforceFieldValues(salesforceSync);
-        setDynamicOptions(data.options || []);
-      } catch (error) {
-        console.error('Failed to fetch Salesforce options:', error);
-        setDynamicOptions([]);
-      } finally {
-        setLoadingDynamicOptions(false);
-      }
-    };
-
-    fetchSalesforceOptions();
-  }, [servar.metadata.salesforce_sync]);
 
   let options;
   if (dynamicOptions.length > 0) {
