+++ src/hooks/useSalesforceSync.ts
@@ -1,0 +1,29 @@
+import { useEffect, useState } from 'react';
+import FeatheryClient from '../utils/featheryClient';
+
+export default function useSalesforceSync(salesforceSync: any) {
+  const [dynamicOptions, setDynamicOptions] = useState<string[]>([]);
+  const [loadingDynamicOptions, setLoadingDynamicOptions] = useState(false);
+
+  useEffect(() => {
+    if (!salesforceSync) return;
+
+    const fetchSalesforceOptions = async () => {
+      setLoadingDynamicOptions(true);
+      try {
+        const client = new FeatheryClient();
+        const data = await client.fetchSalesforceFieldValues(salesforceSync);
+        setDynamicOptions(data.options || []);
+      } catch (error) {
+        console.error('Failed to fetch Salesforce options:', error);
+        setDynamicOptions([]);
+      } finally {
+        setLoadingDynamicOptions(false);
+      }
+    };
+
+    fetchSalesforceOptions();
+  }, [salesforceSync]);
+
+  return { dynamicOptions, loadingDynamicOptions };
+}
